Introduction of JSoup and Shokuhin Logo. Font Size adjustments for Tablet
Used JSoup to remove use of list.hmap, and therefore allow immediate
updating of Recipe List when recipes are added or removed. Allows for
easier upload and deletion of recipes to HTTP server.
Desktop Application needs to be modified to reflect this.

Created Logo, and applied it to the App.

Set font size adjustments to main Viewer, to improve usage on 1600x2560
resolution tablet.